+++ src/header.js
@@ -1,12 +1,7 @@
 /* eslint-disable react/prop-types */
 import React from 'react';
 import defaultStrategy from './default-strategy';
-
-const defaultStyles = {
-  container: {},
-  value: {},
-  order: {}
-};
+import order from './order';
 
 const header = ({
   sortable,
@@ -22,45 +17,31 @@ const header = ({
     throw new Error('header - Missing getSortingColumns!');
   }
 
-  return (value, extra) => {
-    const sortingColumns = getSortingColumns();
-    const sortingColumn = (
-      sortingColumns && sortingColumns[extra[strategy.fieldName]]
-    ) || {};
-    const sortingPosition = sortingColumn.position;
-
-    return (
-      <div
-        className="sort-container"
-        style={defaultStyles.container}
-        {...props.container}
+  return (value, extra) => (
+    <div
+      className="sort-container"
+      {...props.container}
+    >
+      <span
+        className="sort-value"
+        {...props.value}
       >
-        <span
-          className="sort-value"
-          style={defaultStyles.value}
-          {...props.value}
-        >
-          {value}
-        </span>
-        {{}.hasOwnProperty.call(sortingColumn, 'position') ?
-          <span
-            className="sort-order"
-            style={defaultStyles.order}
-            {...props.order}
-          >
-            {sortingPosition + 1}
-          </span> : null
-        }
-        {sortable && React.createElement(
-          'span',
-          sortable(
-            value,
-            extra
-          )
-        )}
-      </div>
-    );
-  };
+        {value}
+      </span>
+      {order({
+        getSortingColumns,
+        props: props.order,
+        strategy
+      })(value, extra)}
+      {sortable ? React.createElement(
+        'span',
+        sortable(
+          value,
+          extra
+        )
+      ) : null}
+    </div>
+  );
 };
 
 export default header;